extracting map sector & frames
