+++ src/Service/Sync/DisplayNameService.php
@@ -26,34 +26,14 @@ class DisplayNameService implements DisplayNameServiceInterface
     }
 
     /**
-     * @param string $fileName
-     *
-     * @return string
-     */
-    private function trimFilename(string $fileName)
-    {
-        // strip file ending
-        $ending = pathinfo($fileName, PATHINFO_EXTENSION);
-
-        // strip duplicate
-        if (preg_match('/_duplicate_[0-9]{4}-[0-9]{2}-[0-9]{2}T[0-9]{2}_[0-9]{2}.' . $ending . '/', $fileName, $matches, PREG_OFFSET_CAPTURE)) {
-            $index = $matches[0][1];
-            $output = mb_substr($fileName, 0, $index);
-        } else {
-            $output = pathinfo($fileName, PATHINFO_FILENAME);
-        }
-
-        return $output;
-    }
-
-    /**
      * @param string $imageName
      *
      * @return string
      */
     public function forConstructionSiteImage(string $imageName)
     {
-        $output = $this->trimFilename($imageName);
+        $output = $this->trimEnding($imageName);
+        $output = $this->trimWhitespace($output);
 
         return $output;
     }
@@ -65,7 +45,7 @@ class DisplayNameService implements DisplayNameServiceInterface
      */
     public function forMapFile(string $mapName)
     {
-        $output = $this->trimFilename($mapName);
+        $output = $this->trimEnding($mapName);
 
         // replace _ with space
         $output = str_replace('_', ' ', $output);
@@ -94,8 +74,9 @@ class DisplayNameService implements DisplayNameServiceInterface
             $output = $before . $after;
         }
 
-        // remove multiple whitespaces
-        return preg_replace('/\s+/', ' ', $output);
+        $output = $this->trimWhitespace($output);
+
+        return $output;
     }
 
     /**
@@ -191,12 +172,119 @@ class DisplayNameService implements DisplayNameServiceInterface
             // join parts back together
             $newName = implode(' ', $mapParts[$counter++]);
 
-            // remove multiple whitespaces
-            $newName = preg_replace('/\s+/', ' ', $newName);
+            $newName = $this->trimWhitespace($newName);
 
             $resultingNames[$key] = $newName;
         }
 
         return $resultingNames;
+    }
+
+    /**
+     * @param string[] $elementNames as an (int id => string name) structure
+     * @param callable $createNewElement called as $addElement(string $name); should return int id of the new element
+     * @param callable $assignChildToParent called with $assignParent(string $childId, string $parentId)
+     */
+    public function putIntoTreeStructure(array $elementNames, callable $createNewElement, callable $assignChildToParent)
+    {
+        // ensure a map exists for all common prefixes (as a folder)
+        $prefixCountMap = $this->createPrefixMap($elementNames);
+
+        // ensure an element exists for all common prefixes
+        foreach ($prefixCountMap as $prefix => $count) {
+            if ($count > 1) {
+                if (!array_key_exists($prefix, $prefixCountMap)) {
+                    $newElementId = $createNewElement($prefix);
+                    $elementNames[$newElementId] = $prefix;
+                }
+            }
+        }
+
+        // create dictionary for each name to point to the first element with that name
+        $prefixElementIdMap = [];
+        foreach ($elementNames as $id => $name) {
+            if (!array_key_exists($name, $prefixElementIdMap)) {
+                $prefixElementIdMap[$name] = $id;
+            }
+        }
+
+        // find longest matching prefix & set as parent
+        foreach ($elementNames as $elementKey => $elementName) {
+            $possibleParentPrefix = $elementName;
+            while (mb_strpos($possibleParentPrefix, ' ') !== false) {
+                // cut off last part of name separated by space
+                $possibleParentPrefix = mb_substr($possibleParentPrefix, 0, mb_strrpos($possibleParentPrefix, ' '));
+
+                // assign to parent if found
+                if (array_key_exists($possibleParentPrefix, $prefixElementIdMap)) {
+                    $assignChildToParent($elementKey, $prefixElementIdMap[$possibleParentPrefix]);
+                }
+            }
+        }
+    }
+
+    /**
+     * counts prefixes occurring and returns a dictionary with (prefix => count).
+     *
+     * @param string[] $names
+     *
+     * @return int[]
+     */
+    private function createPrefixMap(array $names)
+    {
+        $prefixMap = [];
+
+        foreach ($names as $name) {
+            $currentPrefix = $name;
+
+            while (true) {
+                if (!array_key_exists($currentPrefix, $prefixMap)) {
+                    $prefixMap[$currentPrefix] = 1;
+                } else {
+                    ++$prefixMap[$currentPrefix];
+                }
+
+                $newCutoff = \mb_strripos($currentPrefix, ' ');
+                if ($newCutoff === false) {
+                    break;
+                }
+
+                $currentPrefix = trim(\mb_substr($currentPrefix, 0, $newCutoff));
+            }
+        }
+
+        return $prefixMap;
+    }
+
+    /**
+     * @param string $fileName
+     *
+     * @return string
+     */
+    private function trimEnding(string $fileName)
+    {
+        // strip file ending
+        $ending = pathinfo($fileName, PATHINFO_EXTENSION);
+
+        // strip duplicate
+        if (preg_match('/_duplicate_[0-9]{4}-[0-9]{2}-[0-9]{2}T[0-9]{2}_[0-9]{2}.' . $ending . '/', $fileName, $matches, PREG_OFFSET_CAPTURE)) {
+            $index = $matches[0][1];
+            $output = mb_substr($fileName, 0, $index);
+        } else {
+            $output = pathinfo($fileName, PATHINFO_FILENAME);
+        }
+
+        return $output;
+    }
+
+    /**
+     * @param string $text
+     *
+     * @return string
+     */
+    private function trimWhitespace(string $text)
+    {
+        // remove multiple whitespaces
+        return preg_replace('/\s+/', ' ', $text);
     }
 }
