+++ src/Service/Sync/Interfaces/DisplayNameServiceInterface.php
@@ -40,4 +40,11 @@ interface DisplayNameServiceInterface
      * @return string
      */
     public function forConstructionSite(string $folderName);
+
+    /**
+     * @param string[] $elementNames as an (int id => string name) structure
+     * @param callable $createNewElement called as $addElement(string $name); should return int id of the new element
+     * @param callable $assignChildToParent called with $assignParent(string $childId, string $parentId)
+     */
+    public function putIntoTreeStructure(array $elementNames, callable $createNewElement, callable $assignChildToParent);
 }
