+++ src/Service/Sync/MapService.php
@@ -14,13 +14,11 @@ namespace App\Service\Sync;
 use App\Entity\ConstructionSite;
 use App\Entity\Map;
 use App\Entity\MapFile;
-use App\Entity\MapSector;
-use App\Model\Frame;
-use App\Model\Point;
 use App\Model\SyncTransaction;
 use App\Service\Interfaces\PathServiceInterface;
 use App\Service\Sync\Interfaces\DisplayNameServiceInterface;
 use App\Service\Sync\Interfaces\FileServiceInterface;
+use App\Service\Sync\Interfaces\MapFileServiceInterface;
 use App\Service\Sync\Interfaces\MapServiceInterface;
 
 class MapService implements MapServiceInterface
@@ -40,109 +38,17 @@ class MapService implements MapServiceInterface
      */
     private $fileService;
 
-    public function __construct(PathServiceInterface $pathService, DisplayNameServiceInterface $displayNameService, FileServiceInterface $fileService)
+    /**
+     * @var MapFileServiceInterface
+     */
+    private $mapFileService;
+
+    public function __construct(PathServiceInterface $pathService, DisplayNameServiceInterface $displayNameService, FileServiceInterface $fileService, MapFileServiceInterface $mapFileService)
     {
         $this->pathService = $pathService;
         $this->displayNameService = $displayNameService;
         $this->fileService = $fileService;
-    }
-
-    /**
-     * @param SyncTransaction $syncTransaction
-     * @param ConstructionSite $constructionSite
-     * @param Map[] $maps
-     *
-     * @return Map[]
-     */
-    private function createTreeStructure(SyncTransaction $syncTransaction, ConstructionSite $constructionSite, array &$maps)
-    {
-        $mapLookup = [];
-        foreach ($maps as $map) {
-            $mapLookup[$map->getName()] = $map;
-        }
-        $mapNames = array_keys($mapLookup);
-
-        // ensure a map exists for all common prefixes (as a folder)
-        $prefixMap = $this->createPrefixMap(array_keys($mapLookup));
-        foreach ($prefixMap as $prefix => $count) {
-            if ($count > 1) {
-                if (!array_key_exists($prefix, $mapLookup)) {
-                    $map = new Map();
-                    $map->setConstructionSite($constructionSite);
-                    $map->setName($prefix);
-                    $mapLookup[$prefix] = $map;
-                    $maps[] = $map;
-                    $syncTransaction->persist($map);
-                }
-            }
-        }
-
-        // find longest matching prefix & set as parent
-        foreach ($maps as $map) {
-            $parts = explode(' ', $map->getName());
-            for ($i = \count($parts); $i > 0; --$i) {
-                $prefix = '';
-                for ($j = 0; $j < $i; ++$j) {
-                    $prefix .= $parts[$j] . ' ';
-                }
-
-                // find shortest maps name with that prefix
-                $shortestPrefixMatch = null;
-                foreach ($mapNames as $mapName) {
-                    // prefix match
-                    if (\mb_strpos($mapName, $prefix) === 0) {
-                        if ($shortestPrefixMatch === null || \mb_strlen($mapName) < \mb_strlen($shortestPrefixMatch)) {
-                            $shortestPrefixMatch = $mapName;
-                        }
-                    }
-                }
-
-                // if map found, set as parent & stop for current map
-                if ($shortestPrefixMatch !== null) {
-                    $parent = $mapLookup[$shortestPrefixMatch];
-                    if ($parent !== $map) {
-                        $map->setParent($parent);
-                        $syncTransaction->persist($map);
-                    }
-                    break;
-                }
-            }
-        }
-
-        return $maps;
-    }
-
-    /**
-     * counts prefixes occurring and returns a dictionary with (prefix => count).
-     *
-     * @param string[] $names
-     *
-     * @return int[]
-     */
-    private function createPrefixMap(array $names)
-    {
-        $prefixMap = [];
-
-        foreach ($names as $name) {
-            $currentPrefix = $name;
-
-            while (true) {
-                if (!array_key_exists($currentPrefix, $prefixMap)) {
-                    $prefixMap[$currentPrefix] = 1;
-                } else {
-                    ++$prefixMap[$currentPrefix];
-                }
-
-                $newCutoff = \mb_strripos($currentPrefix, ' ');
-                if ($newCutoff === false) {
-                    break;
-                }
-
-                $currentPrefix = trim(\mb_substr($currentPrefix, 0, $newCutoff));
-            }
-        }
-
-        return $prefixMap;
+        $this->mapFileService = $mapFileService;
     }
 
     /**
@@ -151,277 +57,16 @@ class MapService implements MapServiceInterface
      */
     public function syncConstructionSiteMaps(SyncTransaction $syncTransaction, ConstructionSite $constructionSite)
     {
-        // get all files from the directory
-        /** @var MapFile[] $mapFiles */
-        $mapFiles = $constructionSite->getMapFiles()->toArray();
+        $this->mapFileService->syncMapFiles($syncTransaction, $constructionSite);
+
         /** @var Map[] $maps */
         $maps = $constructionSite->getMaps()->toArray();
-
-        $this->findNewMapFiles($syncTransaction, $constructionSite, $mapFiles);
-
-        $this->syncMapFiles($syncTransaction, $constructionSite, $mapFiles);
-
-        $this->refreshMapFileDisplayFileNames($syncTransaction, $mapFiles);
+        /** @var MapFile[] $mapFiles */
+        $mapFiles = $constructionSite->getMapFiles()->toArray();
 
         $this->assignMapFilesToMaps($syncTransaction, $constructionSite, $mapFiles, $maps);
 
         $this->createTreeStructure($syncTransaction, $constructionSite, $maps);
-    }
-
-    /**
-     * @param SyncTransaction $syncTransaction
-     * @param ConstructionSite $constructionSite
-     * @param MapFile[] $mapFiles
-     */
-    private function findNewMapFiles(SyncTransaction $syncTransaction, ConstructionSite $constructionSite, array &$mapFiles)
-    {
-        $mapsDirectory = $this->pathService->getFolderForMapFile($constructionSite);
-        /** @var MapFile[] $newMapFiles */
-        $newMapFiles = $this->fileService->getFiles($mapsDirectory, '.pdf', $mapFiles, function () {
-            return new MapFile();
-        });
-
-        foreach ($newMapFiles as $newMapFile) {
-            $newMapFile->setConstructionSite($constructionSite);
-            $syncTransaction->persist($newMapFile);
-            $mapFiles[] = $newMapFile;
-        }
-    }
-
-    /**
-     * @param SyncTransaction $syncTransaction
-     * @param ConstructionSite $constructionSite
-     * @param MapFile[] $mapFiles
-     */
-    private function syncMapFiles(SyncTransaction $syncTransaction, ConstructionSite $constructionSite, array $mapFiles)
-    {
-        $directory = $this->pathService->getFolderForMapFile($constructionSite);
-        foreach ($mapFiles as $mapFile) {
-            $fileNameWithoutExtension = mb_substr($mapFile->getFilename(), 0, -3);
-
-            $mapSectorsJsonPath = $directory . \DIRECTORY_SEPARATOR . $fileNameWithoutExtension . 'sectors.json';
-            $mapSectors = $this->readMapSectors($mapSectorsJsonPath);
-            $this->syncMapSectors($syncTransaction, $mapFile, $mapSectors);
-
-            $frameJsonPath = $directory . \DIRECTORY_SEPARATOR . $fileNameWithoutExtension . 'sectors.frame.json';
-            $frame = $this->readFrame($frameJsonPath);
-            $this->syncFrame($syncTransaction, $mapFile, $frame);
-        }
-    }
-
-    /**
-     * @param SyncTransaction $syncTransaction
-     * @param MapFile $mapFile
-     * @param Frame|null $frame
-     */
-    private function syncFrame(SyncTransaction $syncTransaction, MapFile $mapFile, ?Frame $frame)
-    {
-        if ($mapFile->getSectorFrame() === null && $frame === null) {
-            return;
-        }
-
-        if ($mapFile->getSectorFrame() !== null && $frame !== null && !$frame->equals($mapFile->getSectorFrame())) {
-            return;
-        }
-
-        $mapFile->setSectorFrame($frame);
-        $syncTransaction->persist($mapFile);
-    }
-
-    /**
-     * @param string $filePath
-     *
-     * @return Frame|null
-     */
-    private function readFrame(string $filePath)
-    {
-        if (!file_exists($filePath)) {
-            return null;
-        }
-
-        $json = json_decode(file_get_contents($filePath));
-
-        if (!property_exists($json, 'upper') ||
-            !property_exists($json, 'lower') ||
-            !property_exists($json, 'left') ||
-            !property_exists($json, 'right')) {
-            return null;
-        }
-
-        $frame = new Frame();
-        $frame->startX = $json->left;
-        $frame->startY = $json->right;
-        $frame->width = $json->right - $json->left;
-        $frame->height = $json->lower - $json->upper;
-
-        return $frame;
-    }
-
-    /**
-     * @param SyncTransaction $syncTransaction
-     * @param MapFile $mapFile
-     * @param MapSector[] $newMapSectors
-     */
-    private function syncMapSectors(SyncTransaction $syncTransaction, MapFile $mapFile, array $newMapSectors)
-    {
-        /** @var MapSector[] $existingSectorsLookup */
-        $existingSectorsLookup = [];
-        foreach ($mapFile->getSectors() as $sector) {
-            $existingSectorsLookup[$sector->getIdentifier()] = $sector;
-        }
-
-        /** @var MapSector[] $newSectorsLookup */
-        $newSectorsLookup = [];
-        foreach ($newMapSectors as $newMapSector) {
-            $newSectorsLookup[$newMapSector->getIdentifier()] = $newMapSector;
-        }
-
-        // remove replaced / not found ones
-        foreach ($existingSectorsLookup as $key => $sector) {
-            if (!array_key_exists($key, $newSectorsLookup) || !$sector->equals($newSectorsLookup[$key])) {
-                $mapFile->getSectors()->remove($sector);
-                $syncTransaction->remove($sector);
-            } else {
-                unset($newSectorsLookup[$key]);
-            }
-        }
-
-        // add new ones
-        foreach ($newMapSectors as $newSector) {
-            $newSector->setMapFile($mapFile);
-            $mapFile->getSectors()->add($newSector);
-            $syncTransaction->persist($newSector);
-        }
-    }
-
-    /**
-     * @param $filePath
-     *
-     * @return MapSector[]|array
-     */
-    private function readMapSectors($filePath)
-    {
-        if (!file_exists($filePath)) {
-            return [];
-        }
-
-        $json = json_decode(file_get_contents($filePath));
-        if (!property_exists($json, 'floor') || !\is_array($json->floor)) {
-            return [];
-        }
-
-        /** @var MapSector[] $mapSectors */
-        $mapSectors = [];
-        foreach ($json->floor as $item) {
-            if (!property_exists($item, 'name') || !property_exists($item, 'points') || !property_exists($item, 'identifier') || !\is_array($item->points)) {
-                continue;
-            }
-
-            $mapSector = new MapSector();
-            $cleanChars = '\t\n\r\0\x0B\'"';
-            $mapSector->setIdentifier(trim($item->identifier, $cleanChars));
-            $mapSector->setName(trim($item->name, $cleanChars));
-            $mapSector->setColor($this->getColorFromMapSectorName($mapSector->getName()));
-            $mapSector->setPoints($this->getPoints($item->points));
-            $mapSectors[] = $mapSector;
-        }
-
-        return $mapSectors;
-    }
-
-    /**
-     * @param string $name
-     *
-     * @return string
-     */
-    private function getColorFromMapSectorName(string $name)
-    {
-        $name = trim(mb_strtolower($name));
-
-        // remove all other chars containing no information
-        $name = preg_replace("/[\s\-&]/", '', $name);
-
-        $kitchen = ['küche', 'essen', 'kochen'];
-        $wetAreas = ['wc', 'bad', 'klo', 'toilette'];
-        $livingArea = ['wohn', 'zimmer', 'essen', 'schlaf', 'gäste', 'gast'];
-        $storage = ['keller', 'auto', 'garage', 'wasch', 'wirtschaft'];
-
-        $checkIfMatch = function (array $needles) use ($name) {
-            foreach ($needles as $needle) {
-                if (\mb_strpos($name, $needle) !== false) {
-                    return true;
-                }
-            }
-
-            return false;
-        };
-
-        if ($checkIfMatch($kitchen)) {
-            // red because of fire
-            return 'FF0000';
-        }
-
-        if ($checkIfMatch($wetAreas)) {
-            // blue because of water
-            return '#0040FF';
-        }
-
-        if ($checkIfMatch($livingArea)) {
-            // green because of nature
-            return '#00FFFF';
-        }
-
-        if ($checkIfMatch($storage)) {
-            // grey because its dark there
-            return '#808080';
-        }
-
-        // default to black
-        return '#000000';
-    }
-
-    /**
-     * @param \stdClass[] $pointsJson
-     *
-     * @return Point[]|array
-     */
-    private function getPoints($pointsJson)
-    {
-        /** @var Point[] $points */
-        $points = [];
-        foreach ($pointsJson as $item) {
-            if (!property_exists($item, 'x') || !property_exists($item, 'y')) {
-                continue;
-            }
-
-            $point = new Point();
-            $point->x = $item->x;
-            $point->y = $item->x;
-            $points[] = $point;
-        }
-
-        return $points;
-    }
-
-    /**
-     * @param SyncTransaction $syncTransaction
-     * @param MapFile[] $mapFiles
-     */
-    private function refreshMapFileDisplayFileNames(SyncTransaction $syncTransaction, array $mapFiles)
-    {
-        $mapNames = [];
-        foreach ($mapFiles as $mapFile) {
-            $mapNames[] = $this->displayNameService->forMapFile($mapFile->getFilename());
-        }
-        $mapNames = $this->displayNameService->normalizeMapNames($mapNames);
-        $counter = 0;
-        foreach ($mapFiles as $mapFile) {
-            $newDisplayFilename = $mapNames[$counter++];
-            if ($newDisplayFilename !== $mapFile->getDisplayFilename()) {
-                $mapFile->setDisplayFilename($newDisplayFilename);
-                $syncTransaction->persist($mapFile);
-            }
-        }
     }
 
     /**
@@ -475,5 +120,46 @@ class MapService implements MapServiceInterface
                 $syncTransaction->persist($map);
             }
         }
+    }
+
+    /**
+     * @param SyncTransaction $syncTransaction
+     * @param ConstructionSite $constructionSite
+     * @param Map[] $maps
+     */
+    private function createTreeStructure(SyncTransaction $syncTransaction, ConstructionSite $constructionSite, array &$maps)
+    {
+        $id = 0;
+        /** @var Map[] $mapLookup */
+        $mapLookup = [];
+        $idNameLookup = [];
+        foreach ($maps as $map) {
+            $idNameLookup[$id] = $map->getName();
+            $mapLookup[$id++] = $map;
+        }
+
+        $assignChildToParent = function ($childId, $parentId) use (&$mapLookup, $syncTransaction) {
+            $childMap = $mapLookup[$childId];
+            $parentMap = $mapLookup[$parentId];
+
+            $childMap->setParent($parentMap);
+            $parentMap->getChildren()->add($childMap);
+
+            $syncTransaction->persist($childMap);
+        };
+
+        $createNewElement = function ($name) use (&$id, &$mapLookup, $syncTransaction, $constructionSite) {
+            $map = new Map();
+            $map->setName($name);
+
+            $map->setConstructionSite($constructionSite);
+            $constructionSite->getMaps()->add($map);
+
+            $mapLookup[$id++] = $map;
+
+            return $id - 1;
+        };
+
+        $this->displayNameService->putIntoTreeStructure($idNameLookup, $createNewElement, $assignChildToParent);
     }
 }
